refactor: move FeatureDocs.jsx into components/
Repo-root JSX component belonged alongside the other UI components.
Updates the import in grimm-reader.jsx, the internal `./features` ->
`../features` path, and the tailwind/guidelines/test references.
